moved duplicated presentToast to alertService
