+++ src/app/pages/emergency/emergency.page.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
 import { HttpHeaders, HttpClient } from '@angular/common/http';
 import { CacheService } from 'ionic-cache';
-import { Platform, ToastController } from '@ionic/angular';
+import { Platform } from '@ionic/angular';
 import { Keyboard } from '@ionic-native/keyboard/ngx';
 import * as jquery from 'jquery';
 import { Contacts, Contact, ContactField, ContactName } from '@ionic-native/contacts/ngx';
@@ -11,6 +11,7 @@ import { ConnectionService } from 'src/app/services/connection/connection.servic
 import { NavigatorService } from 'src/app/services/navigator/navigator.service';
 import { ConfigService } from 'src/app/services/config/config.service';
 import { TranslateService } from '@ngx-translate/core';
+import { AlertService } from 'src/app/services/alert/alert.service';
 
 @Component({
   selector: 'app-emergency',
@@ -36,7 +37,7 @@ export class EmergencyPage implements OnInit {
     // tslint:disable-next-line: deprecation
     private contacts: Contacts,
     private callNumber: CallNumber,
-    private toastCtrl: ToastController,
+    private alert: AlertService,
     private translate: TranslateService
   ) { }
 
@@ -233,7 +234,7 @@ export class EmergencyPage implements OnInit {
         if (!contactFound) {
           if (contactID) { contact.id = contactID; }
           this.saveContact(contact);
-        } else { this.presentToast(this.translate.instant('alert.contact-exists')); }
+        } else { this.alert.presentToast(this.translate.instant('alert.contact-exists')); }
       }, error => {
         console.log('[Error]: While finding contacts...');
         console.log(error);
@@ -246,27 +247,13 @@ export class EmergencyPage implements OnInit {
     contact.save().then(
       () => {
         console.log('Contact saved!', contact);
-        this.presentToast(this.translate.instant('alert.contact-export-success'));
+        this.alert.presentToast(this.translate.instant('alert.contact-export-success'));
       },
       (error: any) => {
         console.error('Error saving contact.', error);
-        this.presentToast(this.translate.instant('alert.contact-export-fail'));
+        this.alert.presentToast(this.translate.instant('alert.contact-export-fail'));
       }
     );
-  }
-
-  /**
-   * @name presentToast
-   * @param message
-   */
-  async presentToast(message) {
-    const toast = await this.toastCtrl.create({
-      message: message,
-      duration: 2000,
-      position: 'top',
-      cssClass: 'toastPosition'
-    });
-    toast.present();
   }
 
   openMail(mail) {
