+++ src/app/pages/practice/detailed-practice.modal.ts
@@ -1,9 +1,10 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { ModalController, ToastController } from '@ionic/angular';
+import { ModalController } from '@ionic/angular';
 import { WebIntentService } from 'src/app/services/web-intent/web-intent.service';
 import { IConfig } from 'src/app/lib/interfaces';
 import { ConfigService } from 'src/app/services/config/config.service';
 import { TranslateService } from '@ngx-translate/core';
+import { AlertService } from 'src/app/services/alert/alert.service';
 
 @Component({
   selector: 'detailed-practice-modal-page',
@@ -19,7 +20,7 @@ export class DetailedPracticeModalPage implements OnInit {
       private modalCtrl: ModalController,
       private webIntent: WebIntentService,
       private translate: TranslateService,
-      private toastCtrl: ToastController
+      private alert: AlertService
     ) {
   }
 
@@ -42,24 +43,10 @@ export class DetailedPracticeModalPage implements OnInit {
     this.isFavorite = !this.isFavorite;
 
     if (!this.isFavorite) {
-      this.presentToast(this.translate.instant('hints.text.favRemoved'));
+      this.alert.presentToast(this.translate.instant('hints.text.favRemoved'));
     } else {
-      this.presentToast(this.translate.instant('hints.text.favAdded'));
+      this.alert.presentToast(this.translate.instant('hints.text.favAdded'));
     }
-  }
-
-  /**
-   * @name presentToast
-   * @param message
-   */
-  async presentToast(message) {
-    const toast = await this.toastCtrl.create({
-      message: message,
-      duration: 2000,
-      position: 'top',
-      cssClass: 'toastPosition'
-    });
-    toast.present();
   }
 
 }
